# src/apistubs/data/specs/ministubs.openapi.yaml
openapi: 3.1.1
info:
  title: |
    [APIStubs]
  version: 1.0.0
  description: |

    TBD
paths:
...
  /{env}/settings/:
    get:
      description: <a target="_blank" href="https://example.net/apidocs/settings/">https://example.net/apidocs/settings/</a>
      tags:
      - 4. StubsSetup
    post:
      description: curl https://example.net/apidocs/settings/ -X POST --data-binary @apistubs.yaml
      requestBody:
        content:
          application/json:
            example:
              account:
                post#/accounts/create/:
                  '409':
                    status: does_not_exist
              takeout:
                get#/requests: 200-not_found
              service:
                post#/personal/takeout/confimation/: 500
                post#/personal/takeout/email/:
                  '200':
                    HEADERS:
                      Etag: c561c68d0ba92bbeb8b0f612a9199f722e3a621a
                    status: custom
            schema: {}
          application/yaml:
            example:
              account:
                post#/accounts/create/:
                  '409':
                    status: does_not_exist
              takeout:
                get#/requests: 200-not_found
              wargamingcom:
                get#/redirect:
                  '301':
                    HEADERS:
                      Location: https://wargaming.com/
                    status: custom
              service:
                post#/personal/takeout/confimation/: 500
                post#/personal/takeout/email/:
                  '200':
                    HEADERS:
                      Etag: c561c68d0ba92bbeb8b0f612a9199f722e3a621a
                    status: custom
            schema: {}
        required: true
      tags:
      - 4. StubsSetup
    patch:
      description: curl https://example.net/apidocs/settings/ -X PATCH --data-binary @apistubs.yaml
      tags:
      - 4. StubsSetup
    delete:
      tags:
      - 4. StubsSetup
...
servers:
...